print norms

+++ test/test_libs.py
@@ -1158,7 +1158,7 @@ class TestD4RL:
         keys = set(data_from_env._storage._storage.keys(True, True))
         keys = keys.intersection(data_true._storage._storage.keys(True, True))
         for key in keys:
-            print(key, (data_true._storage._storage[key] - data_from_env._storage._storage[key]).norm())
+            print(key, (data_true._storage._storage[key].float() - data_from_env._storage._storage[key].float()).norm())
         # assert_allclose_td(
         #     data_true._storage._storage.select(*keys),
         #     data_from_env._storage._storage.select(*keys),
